move expect_fun_call to its own module
